This starts the 2.0.0 release development.
Added examples to all modules, icluding the README and the main module,
with extra information. This closes #18. Also changed return types to
enable easier use of the library (no need to remember tuple meanings).
This replaces the VSOP to keplerian function for a simple From/Into
trait. It's a breaking change, and it closes #19.