Fix build on non-x86 platforms.
Account for non-intel architectures when compiling the dispatch
to avx or fallback implementations. The cfg attribute on the
call site must be the same as the attribute governing compilation
of the function definition, and the fallback must be enabled for
the opposite condition.

Previously the fallback was only enabled for the no_std case.

Restores positive `cargo +nightly test` run on my aarch64 laptop.